upd test files, add issue references
