+++ e2e/tests/popupEscHandling-example.tsx
@@ -5,6 +5,8 @@ const inputItems = ['Bruce Wayne', 'Natasha Romanoff', 'Steven Strange']
 
 export const selectors = {
   popupTriggerId: 'trigger',
+  popupContentClass: Popup.slotClassNames.content,
+  dropdownTriggerClass: Dropdown.slotClassNames.triggerButton,
 }
 
 const PopupExample = () => (
