+++ e2e/tests/popupEscHandling-test.ts
@@ -1,0 +1,21 @@
+import { selectors } from './popupEscHandling-example'
+
+const popupTrigger = `#${selectors.popupTriggerId}`
+const popupContent = `.${selectors.popupContentClass}`
+const dropdownTriggerButton = `.${selectors.dropdownTriggerClass}`
+
+// https://github.com/stardust-ui/react/issues/1079
+describe('Popup - on ESC key press', () => {
+  beforeEach(async () => {
+    await e2e.gotoTestCase(__filename, popupTrigger)
+  })
+
+  it('is not closed if ESC is handled by children', async () => {
+    await e2e.clickOn(popupTrigger)
+
+    await e2e.clickOn(dropdownTriggerButton) // opens dropdown list
+    await e2e.pressKey('Escape') // closes dropdown list
+
+    expect(await e2e.count(popupContent)).toBeGreaterThan(0)
+  })
+})
